Converting my custom middleware to Passport. Previously, the API route has custom middle that use the jsonwebtoken module to retrieve the token from either the body, query, or header and verify the token and either proceed or stop.
The new code uses passport-jwt to do essentially the same thing but with Passport.  The passport.use() demonstrates a simple verification of the JWT as well as further authorization code (here just verifying the user against an array of users). In the /users/ handler I then demonstrate retrieving the JWT payload to retrieve information about the logged-in user, which is stored at req.user. Simple!