Bug fixes and changed package settings.
Made a better build filename. Changed bower to be a required global install, which is reasonable.

Eliminated usage of the config package since it presently is incompatible with Electron.

Fixed a bug where a silent JS error would be thrown if a done pic is used instead of a video.